Refactor package name
패키지명 ssr을 web으로 변경
