+++ src/test/java/com/seungki/urlshortener/controller/TestUrlValidation.java
@@ -4,7 +4,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.view;
 
-import com.seungki.urlshortener.ssr.service.UrlShortenService;
+import com.seungki.urlshortener.web.service.UrlShortenService;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
